fixed some stuff

+++ wadtek.cpp
@@ -55,21 +55,21 @@ wadtek::file::file(const std::string &path)
     : _handle(new mio::basic_mmap<mio::access_mode::read, uint8_t>(path))
 {
     size_t offset = 0;
-    uint32_t lump_count = 0;
-    uint32_t lump_offset = 0;
+    int32_t lump_count = 0;
+    int32_t lump_offset = 0;
     const uint8_t *data = get_handle(_handle)->data();
     const uint8_t *lump_marker;
     _identification = wadtek::read_string(data, 4);
     offset += 4;
-    lump_count = wadtek::read_int<uint32_t>(data + offset);
+    lump_count = wadtek::read_int<int32_t>(data + offset);
     offset += 4;
-    lump_offset = wadtek::read_int<uint32_t>(data + offset);
+    lump_offset = wadtek::read_int<int32_t>(data + offset);
     lump_marker = data + lump_offset;
     _lumps.reserve(lump_count);
     while(_lumps.size() < _lumps.capacity())
     {
-        offset = wadtek::read_int<uint32_t>(lump_marker);
-        size_t size = wadtek::read_int<uint32_t>(lump_marker + 8);
+        offset = static_cast<size_t>(wadtek::read_int<int32_t>(lump_marker));
+        size_t size = static_cast<size_t>(wadtek::read_int<int32_t>(lump_marker + 8));
         std::string name = wadtek::read_string(lump_marker + 12, 8);
         _lumps.emplace_back(name, data + offset, size);
         lump_marker += 16;
